phase-03-networking-compose/requirements.txt
flask==3.0.3
psycopg2-binary==2.9.9

